feat: proper Mac GPU acceleration — MPS diarization, int8 CPU, MLX fast path
Device detection now handles all platforms correctly:

  CUDA (Windows/Linux + NVIDIA):
    whisper_device=cuda, diarize_device=cuda, compute_type=float16

  Apple Silicon Mac (MPS available, no CUDA):
    whisper_device=cpu   — ctranslate2 does not support Metal/MPS
    diarize_device=mps   — pyannote IS pure PyTorch, uses Metal GPU
    compute_type=int8    — 2-4x faster than float32 on Apple Silicon CPU

  Intel Mac / CPU-only:
    whisper_device=cpu, diarize_device=cpu, compute_type=int8

MLX fast path (optional, pip install whisperapp[mac-gpu]):
  If mlx-whisper is installed on an Apple Silicon Mac, Whisper transcription
  runs via Apple's MLX framework on the Neural Engine / GPU — significantly
  faster than ctranslate2 on CPU. Falls back gracefully if not installed.

Other changes:
  - _cleanup_memory(): clears MPS buffer pool via torch.mps.empty_cache()
  - streaming.py: auto device inherits from worker._WHISPER_DEVICE/_COMPUTE_TYPE
  - streaming.py polish(): uses _DIARIZE_DEVICE for diarization pipeline
  - server.py: /info endpoint exposes device selection and acceleration mode
  - cli.py: `whisperapp info` command shows device/acceleration summary
  - pyproject.toml: mac-gpu optional extra (mlx-whisper)
  - test_streaming: update CPU compute_type expectation to int8

135 passing, 3 pre-existing streaming integration failures unchanged.

+++ tests/test_streaming.py
@@ -388,7 +388,8 @@ class TestDeviceAutoDetection:
         from whisperapp.streaming import StreamingEngine
         engine = StreamingEngine(device="auto", compute_type="auto")
         assert engine.device == "cpu"
-        assert engine.compute_type == "float32"
+        # CPU auto-detection now uses int8 (faster than float32 on modern CPUs)
+        assert engine.compute_type == "int8"
 
     def test_streaming_engine_explicit_device_respected(self):
         from whisperapp.streaming import StreamingEngine
